fix: moved get owners details in getCodeOwners
